Fix build on 64-bit systems.
Needed to cast some things to integers for logging code.

+++ xorp/libxipc/xrl_pf_stcp.cc
@@ -1067,7 +1067,7 @@ XrlPFSTCPSender::read_event(BufferedAsyncReader* reader,
     if (xap) {
 	if (xrl_trace.on()) {
 	    XLOG_INFO("rcv, bytes-remaining: %i  xrl: %s\n",
-		      reader->available_bytes(), xap->str().c_str());
+		      (int)(reader->available_bytes()), xap->str().c_str());
 	}
 
 	// Dispatch Xrl and exit
